requirements.txt
asgiref==3.2.7
astroid==2.4.0
dj-database-url==0.5.0
Django==3.0.5
django-heroku==0.3.1
djangorestframework==3.11.0
isort==4.3.21
lazy-object-proxy==1.4.3
Markdown==3.2.1
mccabe==0.6.1
psycopg2==2.7.5
psycopg2-binary==2.8.5
pylint==2.5.0
pytz==2020.1
six==1.14.0
sqlparse==0.3.1
toml==0.10.0
typed-ast==1.4.1
whitenoise==5.0.1
wrapt==1.12.1